+++ directx/Constants.h
@@ -13,6 +13,6 @@ const Vector4 CLEAR = Vector4(1, 1, 1, 0.1f);
 
 const int ORTHOGRAPHIC = 0;
 const float GRID_INTERVAL = 0.1f;
-const int GRIDPOINTS_COUNT = ((2 / 0.1f)) * 4;
+const int GRIDPOINTS_COUNT = (int)(((2 / 0.1f)) * 4);
 
 static const int NUM_SAMPLES = 10;
